split rollback registry import tests
